server/server/requirements.txt
Django==2.0.3
channels==2.0.2
channels-redis==2.1.1
djangorestframework==3.7.7